refactor : [step4] 도메인 패키지 분리로 인한 테스트 코드 패키지 분리
